+++ launcher/pages/global/AccountListPage.h
@@ -60,7 +60,8 @@ public:
     }
 
 public slots:
-    void on_actionAdd_triggered();
+    void on_actionAddMojang_triggered();
+    void on_actionAddMicrosoft_triggered();
     void on_actionRemove_triggered();
     void on_actionSetDefault_triggered();
     void on_actionNoDefault_triggered();
@@ -74,7 +75,6 @@ public slots:
 
 protected slots:
     void ShowContextMenu(const QPoint &pos);
-    void addAccount(const QString& errMsg="");
 
 private:
     void changeEvent(QEvent * event) override;
