+++ launcher/pages/global/AccountListPage.ui
@@ -36,7 +36,8 @@
    <attribute name="toolBarBreak">
     <bool>false</bool>
    </attribute>
-   <addaction name="actionAdd"/>
+   <addaction name="actionAddMicrosoft"/>
+   <addaction name="actionAddMojang"/>
    <addaction name="actionRemove"/>
    <addaction name="actionSetDefault"/>
    <addaction name="actionNoDefault"/>
@@ -44,9 +45,9 @@
    <addaction name="actionUploadSkin"/>
    <addaction name="actionDeleteSkin"/>
   </widget>
-  <action name="actionAdd">
+  <action name="actionAddMojang">
    <property name="text">
-    <string>Add</string>
+    <string>Add Mojang</string>
    </property>
   </action>
   <action name="actionRemove">
@@ -78,6 +79,11 @@
    </property>
    <property name="toolTip">
     <string>Delete the currently active skin and go back to the default one</string>
+   </property>
+  </action>
+  <action name="actionAddMicrosoft">
+   <property name="text">
+    <string>Add Microsoft</string>
    </property>
   </action>
  </widget>
